requirements.txt
djangorestframework==3.9.0
django>=2.2
Pygments==2.3.1
coreapi==2.3.3
psycopg2==2.7.6.1
python-decouple==3.1
gunicorn==19.9.0
